Moved the calendar widget to , set the calendar to emit a signal when the buttons are pressed, and resized the toolpane on the right
