+++ MultiHex/clock.py
@@ -1,7 +1,14 @@
-# this class will be used to keep track of time
+"""
+Three classes are defined here
 
-# it's a perfect clockwork universe. Hurray
+    Time  -  an instance in time. Has a number of years, months, days, etc...
+    Clock - keeps track of time across a whole planet. Incl timezones. Includes utilities for skipping around to significant times (sunrise, sunset). Can return light levels at any latitude/longitude.
+                keeps track of moon too.
+    MultiHexCalendar - A Qt widget for the calendar in the time system described here 
+"""
 
+
+from PyQt5 import QtWidgets, QtCore
 
 from math import pi, floor
 from math import cos, sin, sqrt
@@ -12,29 +19,30 @@ degrees = pi/180.
 
 minutes_in_hour = 60
 hours_in_day    = 24
-days_in_month   = 35
+days_in_month   = 33
 months_in_year  = 12
 
 minutes_in_day   = minutes_in_hour*hours_in_day
 minutes_in_month = minutes_in_day*days_in_month
 minutes_in_year  = minutes_in_month*months_in_year
 
+# Idea: load these days in from a config file! 
 month_list = [ "January",
-            "February",
-            "March",
-            "Aprtil",
-            "May",
-            "June",
-            "July",
-            "August",
-            "September",
-            "October",
-            "November",
-            "December" ]
+            "Febuary",
+            "Maruary",
+            "Apruary",
+            "Mayuary",
+            "Junuary",
+            "Juluary",
+            "Auguary",
+            "Sepuary",
+            "Octuary",
+            "Novuary",
+            "Decuary" ]
 assert( len(month_list) == months_in_year )
 
 
-day_list = [ "Morndas",
+day_list = [ "Monday",
           "Tuesday",
           "Wednesday",
           "Thursday",
@@ -106,7 +114,13 @@ class Time:
         Returns the day of the week given the current time
         """
         n_days = self._day_shift + self.day + days_in_month*(self.month + months_in_year*self.year)
-        return( day_list[ n_days % len(day_list) ])
+        return(  n_days % len(day_list) )
+
+    def get_day_of_week_str(self):
+        """
+        Returns the day of the week as a string. 
+        """
+        return(day_list[self.get_day_of_week()])
 
     def month_str(self):
         """
@@ -670,4 +684,203 @@ class Clock:
 
     def __str__(self):
         return("{}\nThe moon is {}".format(self.get_local_time(0), self.get_moon_phase()) )
+
+# utility for signaling 
+class Signaler(QtCore.QObject):
+    signal = QtCore.pyqtSignal(str)
+
+class MultiHexCalendar(QtWidgets.QWidget):
+    """
+    This implements a QtWidget for the MultiHex Calendar.
+    It creates a little standard grid of days (buttons) below a month and a year switcher. 
+
+    As it is, this doesn't really *do* anything. Later, I'll set it up so that this emits a signal when a day-button is pressed. The signal it emits should include the day the button has on it
+    """
+    def __init__(self, parent, time):
+        """
+        Initializes the widget. 
+
+        The arg 'parent' is the widget's parent
+        The arg 'time' is the starting time for the calendar. 
+        """
+        QtWidgets.QWidget.__init__(self, parent)
+        if not isinstance(time, Time):
+            raise TypeError("Expected {}, got {}".format(Time, type(time)))
+
+        self.time = time
+
+        self.setObjectName("MuliHexCalendar")
+        self.layout = QtWidgets.QVBoxLayout(self)
+
+        # define and add the year label and add the buttons beside it
+        # TODO: (maybe) swap out the year label for something allowing easier large-scale year changes. 
+        self.year_layout = QtWidgets.QHBoxLayout(self)
+        self.leftButton = QtWidgets.QPushButton(self)
+        self.leftButton.setObjectName("leftButton")
+        self.leftButton.setFixedSize(25,25)
+        self.leftButton.setText("<-")
+        self.rightButton = QtWidgets.QPushButton(self)
+        self.rightButton.setObjectName("rightButton")
+        self.rightButton.setFixedSize(25,25)
+        self.rightButton.setText("->")
+        self.yearLabel = QtWidgets.QLabel(self)
+        self.yearLabel.setObjectName("yearLabel")
+        self.yearLabel.setText("{}".format(self.time.year))
+        self.yearLabel.setAlignment(QtCore.Qt.AlignCenter)
+        self.year_layout.addWidget(self.leftButton)
+        self.year_layout.addWidget(self.yearLabel)
+        self.year_layout.addWidget(self.rightButton)
+        self.layout.addItem(self.year_layout)
+
+        # define and add the month combo box and the buttons beside it
+        self.month_layout = QtWidgets.QHBoxLayout(self)
+        self.leftButtonMon = QtWidgets.QPushButton(self)
+        self.leftButtonMon.setObjectName("leftButtonMon")
+        self.leftButtonMon.setFixedSize(25,25)
+        self.leftButtonMon.setText("<-")
+        self.rightButtonMon = QtWidgets.QPushButton(self)
+        self.rightButtonMon.setObjectName("rightButtonMon")
+        self.rightButtonMon.setFixedSize(25,25)
+        self.rightButtonMon.setText("->")
+        self.month_combo = QtWidgets.QComboBox(self)
+        self.month_combo.setObjectName("month_combo")
+        for month in month_list:
+            self.month_combo.addItem(month)
+        self.month_layout.addWidget(self.leftButtonMon)
+        self.month_layout.addWidget(self.month_combo)
+        self.month_layout.addWidget(self.rightButtonMon)
+        self.layout.addItem(self.month_layout)
+        self.month_combo.setCurrentIndex(self.time.month)
+
+        # Add in the labels for hte weekdays
+        weekday_lbls ={}
+        self.weekday_list = QtWidgets.QHBoxLayout(self)
+        for day in range(len(day_list)):
+            weekday_lbls[day] = QtWidgets.QLabel(self)
+            weekday_lbls[day].setText(day_list[day][:2])
+            weekday_lbls[day].setObjectName(day_list[day])
+            weekday_lbls[day].setAlignment(QtCore.Qt.AlignCenter)
+
+            self.weekday_list.addWidget(weekday_lbls[day])
+        self.layout.addItem(self.weekday_list)
+
+        # we put all the buttons in place, and disable the ones that aren't needed
+        # This way we don't have to add/remove buttons as necessary. Just enable/disable them and change the text 
+        self.day_buttons = {}
+        self.calendarGrid =  QtWidgets.QGridLayout(self)
+        self.total_rows = int(days_in_month /len(day_list) ) + 1 #  make sure we have enough rows of buttons
+        row = 0
+        column = 0
+        day = 0 # just a button counter...
+        while row<=self.total_rows:
+            self.day_buttons[day] = QtWidgets.QPushButton(self)
+            self.day_buttons[day].setFixedSize(25,25)
+            self.day_buttons[day].setText("")
+            self.day_buttons[day].setEnabled(False)
+
+            # this is weird, but is a little necessary necessary for scoping...
+            # the function takes a day, and returns a function that accesses that day's text. We then connect the returned function to the button 
+            def get_func(which):
+                def funcy():
+                    temp = self.day_buttons[which]
+                    return( self.buttonPress(temp.text()) )
+                return(funcy)
+            # lambda(x : self.buttonPress( self.day_buttons[x]))
+            self.day_buttons[day].clicked.connect( get_func(day) )
+            #self.day_buttons[day].clicked.connect( lambda x : self.buttonPress( self.day_buttons[x]) )
+            self.calendarGrid.addWidget(self.day_buttons[day],row,column)
+
+            column+=1
+            day+=1
+            if column==(len(day_list)):
+                column = 0
+                row += 1
+        self.days = day
+        self.layout.addItem(self.calendarGrid)
+
+        # Use the current time to choose which buttons to enable
+        self.fill_days()
+
+        self.month_combo.currentIndexChanged.connect(self.change_month)
+        self.leftButton.clicked.connect(self.remove_year)
+        self.rightButton.clicked.connect(self.add_year)
+        self.leftButtonMon.clicked.connect(self.leftMon)
+        self.rightButtonMon.clicked.connect(self.rightMon)
+
+        # create a signal which is emitted when a button is pressed
+        # this is the only way I know how to set up a emit - receive signal system. Used the same technique from the MultiThreading 
+        # so it'll go //calendar.signals.signal.connect( function )//
+        self.signals = Signaler()
+
+    def fill_days(self):
+        """
+        This updates the buttons to the current year/month
+
+        It scrolls over them until we get to the first day of the month: disabling buttons as it goes.
+        It continues along, enabling buttons until it reaches the number of days in the month.
+        Then carries on until all the rest of buttons are off. 
+        """
+        # total days is N
+
+        counting = False
+        days_so_far = 0
+        day = 0 
+        weekday_of_first_of_month = Time(year=self.time.year, month = self.time.month, day=0).get_day_of_week()
+
+        while day < self.days:
+            if day==weekday_of_first_of_month and days_so_far==0:
+                counting = True
+
+            if counting:
+                days_so_far += 1
+                self.day_buttons[day].setText(str(days_so_far))
+                self.day_buttons[day].setEnabled(True)
+            else:
+                self.day_buttons[day].setText("")
+                self.day_buttons[day].setEnabled(False)
+
+            if days_so_far==days_in_month:
+                counting=False
+
+            day+=1
+
+    def buttonPress(self, what):
+        self.signals.signal.emit(what)
+
+
+    def leftMon(self):
+        new = self.month_combo.currentIndex() -1
+        if new<0:
+            self.time = self.time - Time(year=1)
+            self.month_combo.setCurrentIndex(months_in_year-1)
+        else:
+            self.month_combo.setCurrentIndex(new)
+
+    def rightMon(self):
+        new = self.month_combo.currentIndex() +1
+        if new==months_in_year:
+            self.time = self.time + Time(year=1)
+            self.month_combo.setCurrentIndex(0)
+        else:
+            self.month_combo.setCurrentIndex(new)
+
+    def change_month(self):
+        month_diff = self.month_combo.currentIndex() - self.time.month
+        if month_diff > 0 :
+            self.time =self.time+ Time(month=month_diff )
+        elif month_diff <0:
+            self.time = self.time- Time(month=-1*month_diff)
+        self.fill_days()
+        self.yearLabel.setText("{}".format(self.time.year))
+
+    def add_year(self):
+        self.time += Time(year=1)
+        self.yearLabel.setText("{}".format(self.time.year))
+        self.fill_days()
+
+    def remove_year(self):
+        self.time -= Time(year=1)
+        self.yearLabel.setText("{}".format(self.time.year))
+        self.fill_days()
+
